[azure][fix] Warn and ignore HttpResponseError
